# Conflicts:
#	CMakeLists.txt
#	Modules/master_machine/master_process.h
#	Modules/motor/DJImotor/dji_motor.c
#	UserApp/components/shoot/shoot_standard/shoot.c
#	UserApp/os_task.c
#	UserApp/robot/debug/robot.c
#	UserApp/robot/debug/robot.cmake
#	UserApp/robot/hero_mecanum/robot_config.h
#	UserApp/robot/infantry_mecanum/robot_config.h

+++ Modules/motor/motor_def.h
@@ -80,8 +80,8 @@ typedef struct {
   PIDInstance speed_PID;
   PIDInstance angle_PID;
 
-  float pid_ref;         // 将会作为每个环的输入和输出顺次通过串级闭环
-  int16_t final_output;  // pid计算最后的输出，或者说最内环pid计算的输出
+  float pid_ref;       // 将会作为每个环的输入和输出顺次通过串级闭环
+  float final_output;  // pid计算最后的输出，或者说最内环pid计算的输出
 } Motor_Controller_s;
 
 /* 电机类型枚举 */
